Print args before validation, update AllTestFiltered ios
