+++ test_runner/src/main/kotlin/ftl/args/ValidateIosArgs.kt
@@ -3,7 +3,10 @@ package ftl.args
 import ftl.args.yml.Type
 import ftl.ios.IosCatalog
 import ftl.ios.IosCatalog.getSupportedVersionId
+import ftl.ios.xctest.common.mapToRegex
+import ftl.log.logLn
 import ftl.run.exception.FlankConfigurationError
+import ftl.run.exception.FlankGeneralError
 import ftl.run.exception.IncompatibleTestDimensionError
 
 fun IosArgs.validate() = apply {
@@ -17,6 +20,7 @@ fun IosArgs.validate() = apply {
     assertAdditionalIpas()
     validType()
     assertGameloop()
+    assertXcTestRunData()
 }
 
 private fun IosArgs.assertGameloop() {
@@ -83,4 +87,32 @@ private fun IosArgs.validType() {
     val validIosTypes = arrayOf(Type.GAMELOOP, Type.XCTEST)
     if (commonArgs.type !in validIosTypes)
         throw FlankConfigurationError("Type should be one of ${validIosTypes.joinToString(",")}")
+}
+
+private fun IosArgs.assertXcTestRunData() {
+    if (!disableSharding && testTargets.isNotEmpty()) {
+        val filteredMethods = xcTestRunData
+            .shardTargets.values
+            .flatten()
+            .flatMap { it.values }
+            .flatten()
+
+        if (filteredMethods.isEmpty()) throw FlankGeneralError(
+            "Empty shards. Cannot match any method to $testTargets"
+        )
+
+        if (filteredMethods.size < testTargets.size) {
+            val regexList = testTargets.mapToRegex()
+
+            val notMatched = testTargets.filter {
+                filteredMethods.all { method ->
+                    regexList.any { regex ->
+                        regex.matches(method)
+                    }
+                }
+            }
+
+            logLn("WARNING: cannot match test_targets: $notMatched")
+        }
+    }
 }
